// liwords-ui/src/gen/api/vendor/macondo/macondo_pb.d.ts
export declare enum BotRequest_BotCode {
  /**
   * @generated from enum value: HASTY_BOT = 0;
   */
  HASTY_BOT = 0,

  /**
   * @generated from enum value: LEVEL1_CEL_BOT = 1;
   */
  LEVEL1_CEL_BOT = 1,

  /**
   * @generated from enum value: LEVEL2_CEL_BOT = 2;
   */
  LEVEL2_CEL_BOT = 2,

  /**
   * @generated from enum value: LEVEL3_CEL_BOT = 3;
   */
  LEVEL3_CEL_BOT = 3,

  /**
   * @generated from enum value: LEVEL4_CEL_BOT = 4;
   */
  LEVEL4_CEL_BOT = 4,

  /**
   * @generated from enum value: LEVEL1_PROBABILISTIC = 5;
   */
  LEVEL1_PROBABILISTIC = 5,

  /**
   * @generated from enum value: LEVEL2_PROBABILISTIC = 6;
   */
  LEVEL2_PROBABILISTIC = 6,

  /**
   * @generated from enum value: LEVEL3_PROBABILISTIC = 7;
   */
  LEVEL3_PROBABILISTIC = 7,

  /**
   * @generated from enum value: LEVEL4_PROBABILISTIC = 8;
   */
  LEVEL4_PROBABILISTIC = 8,

  /**
   * @generated from enum value: LEVEL5_PROBABILISTIC = 9;
   */
  LEVEL5_PROBABILISTIC = 9,

  /**
   * @generated from enum value: NO_LEAVE_BOT = 10;
   */
  NO_LEAVE_BOT = 10,

  /**
   * @generated from enum value: SIMMING_BOT = 11;
   */
  SIMMING_BOT = 11,

  /**
   * @generated from enum value: HASTY_PLUS_ENDGAME_BOT = 12;
   */
  HASTY_PLUS_ENDGAME_BOT = 12,

  /**
   * harder bots here? inference? ml, etc.
   *
   * @generated from enum value: SIMMING_INFER_BOT = 13;
   */
  SIMMING_INFER_BOT = 13,

  /**
   * This should have probably been number 0.
   *
   * @generated from enum value: UNKNOWN = 100;
   */
  UNKNOWN = 100,
}